fix: separate authOptions to auth.config.js for Next.js App Router compliance
- Create auth.config.js with authOptions configuration
- Update route.js to only export GET and POST handlers
- Fix all import paths to use auth.config.js instead of route.js
- Resolves Next.js build error about invalid Route export field
