+++ app/(dashboard)/(student)/student-settings/page.js
@@ -1,5 +1,5 @@
 import { getServerSession } from 'next-auth/next';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/api/auth/[...nextauth]/auth.config';
 import { getUserProfile } from '@/app/lib/actions/getUserProfile';
 import { redirect } from 'next/navigation';
 import BackToTop from '@/app/backToTop';
