+++ app/(dashboard)/dashboard/page.js
@@ -1,5 +1,5 @@
 import { getServerSession } from 'next-auth/next';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/api/auth/[...nextauth]/auth.config';
 import { redirect } from 'next/navigation';
 
 // 컴포넌트 import
