+++ app/api/auth/phone/send-otp/route.js
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth/next';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/api/auth/[...nextauth]/auth.config';
 import { createClient } from '@supabase/supabase-js';
 import twilio from 'twilio';
 import {
