+++ app/api/user/profile/route.js
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth/next';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/api/auth/[...nextauth]/auth.config';
 import { createClient } from '@supabase/supabase-js';
 
 // Supabase 클라이언트 생성
