+++ app/certificate/[id]/page.js
@@ -1,6 +1,6 @@
 import { getCertificate } from "@/app/lib/certificate/actions/certificateActions";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/api/auth/[...nextauth]/auth.config";
 import { redirect } from 'next/navigation';
 import CertificateView from "@/components/Certificate/CertificateView";
