+++ app/lib/actions/lessonActions.js
@@ -2,7 +2,7 @@
 
 import { supabase } from '@/app/lib/supabase/client';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/api/auth/[...nextauth]/auth.config';
 
 // Create a new lesson
 export async function createLesson(lessonData) {
